+++ shelkovnikov.dmitrii/I2/countorderedelements.cpp
@@ -1,6 +1,6 @@
 #include "countorderedelements.h"
 #include <cstddef>
-unsigned int count_ordered_elements(int *arr, size_t size)
+size_t count_ordered_elements(int *arr, size_t size)
 {
   size_t count = 0;
   for (size_t i = 1; i < size; i++)
